+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ8F9CMGZT17KZ9F31HF990P"
+        "01KJA0G6E7KQ4CMRKFMW84EQMW"
     )
     r"""The unique identifier of the data source"""