+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ8F9CME5RSE2HRJZQAGEXR2"
+        "01KJA0G6E6DT0QQWTAND9J9WXZ"
     )
     r"""The unique identifier of the data source"""